requirements.txt
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.0.0
samv2==0.0.4
scikit-image==0.24.0
supervision==0.24.0
tqdm==4.66.5
ultralytics==8.3.23
ultralytics-thop==2.0.9